Changed 'trip' model name in slot to 'item', cause objects can also be rides. (#273)
Added grouping for rides in the same component
Change the looking up for rides. Now passing id instead of index. Did not have to give the rideId as a prop because the whole ride is being passed as a property to the <ride-card>. Just had to change the parameter of the onRideSelected.
Also changed redundant parameter when sending user to the rideDetailsPage.